python-BE/requirements.txt
fastapi==0.95.1
uvicorn==0.23.1
psycopg2-binary==2.9.5  # Use this for connecting to PostgreSQL
unstructured==0.0.2
unstructured-inference==0.0.1
langchain==0.0.1
llama-index==0.3.8
autogen==0.1.1
crewai==0.1.1
elasticsearch==8.10.1
boto3==1.24.12
redis==4.5.3
python-dotenv==1.0.0
pydantic==1.10.2
unstructured-inference==0.0.1
torch==2.5.1
passlib==1.7.4
sqlalchemy==1.4.23
python-dotenv==0.19.0
boto3==1.18.50
awscli==1.19.50
python-multipart 